+++ OpenGL/Source/c_Npc.cpp
@@ -713,7 +713,7 @@ void c_Npc::RenderNpc()
 		dataManager->getLeaderBoards(data, name);
 
 		RenderTextOnScreen(meshList[TEXT], "LeaderBoard", Color(1, 0, 0), 3, 9, 18);
-		for (int i = 0; i < data.size(); i++)
+		for (int i = 0; i < (int)data.size(); i++)
 		{
 			RenderTextOnScreen(meshList[TEXT], name[i], Color(0, 0, 1), 3, 9, counter);
 			RenderTextOnScreen(meshList[TEXT], std::to_string(data[i]), Color(0, 1, 0), 3, 19, counter);
