+++ OpenGL/Source/c_SecondCar.cpp
@@ -12,10 +12,10 @@ c_SecondCar::c_SecondCar()
 	pos.y = 1;
 	pos.z = 0;
 
-	MaxSpeed = 0.6;
+	MaxSpeed = 0.6f;
 	SteeringAngle = 0;
-	MaxAcceleration = 0.8;
-	Friction = 0.04;
+	MaxAcceleration = 0.8f;
+	Friction = 0.04f;
 	Steering = 4;
 	Driving = false;
 	Backwards = false;
@@ -43,7 +43,7 @@ c_SecondCar::~c_SecondCar()
 void c_SecondCar::Ability(double dt)
 {
 	c_Sound* Audio = c_Sound::getInstance();
-	elapsedTime += dt;
+	elapsedTime += (float)(dt);
 	if (uniqueName == "player2")
 	{
 		if (Application::IsKeyPressed('P'))
@@ -78,7 +78,7 @@ void c_SecondCar::isOffRoad()
 {
 	c_OffRoadManager* manager = c_OffRoadManager::getInstance();
 
-	for (int i = 0; i < manager->getList().size(); i++)
+	for (int i = 0; i < (int)manager->getList().size(); i++)
 	{
 		if (gotCollide(manager->getList()[i], false) || !gotCollide("track", false))
 		{
